change addr type from u32 to usize

+++ src/memory.rs
@@ -1,8 +1,8 @@
 pub trait Memory {
-    fn read_inst(&self, addr: u32) -> u32;
+    fn read_inst(&self, addr: usize) -> u32;
 
-    fn read_word(&self, addr: u32) -> u32;
-    fn write_word(&mut self, addr: u32, data: u32);
+    fn read_word(&self, addr: usize) -> u32;
+    fn write_word(&mut self, addr: usize, data: u32);
 }
 
 #[derive(Debug)]
@@ -11,7 +11,7 @@ pub struct VectorMemory {
 }
 
 impl VectorMemory {
-    pub fn new(size: u32) -> Self {
+    pub fn new(size: usize) -> Self {
         let mut memory = Vec::with_capacity(size);
         memory.resize(size, 0);
 
@@ -21,7 +21,7 @@ impl VectorMemory {
     }
 
     /// read big-endian word located at *addr*
-    fn read_bw(&self, addr: u32) -> u32 {
+    fn read_bw(&self, addr: usize) -> u32 {
         (self.memory[addr] as u32) << 24 |
         (self.memory[addr+1] as u32) << 16 |
         (self.memory[addr+2] as u32) << 8 |
@@ -29,7 +29,7 @@ impl VectorMemory {
     }
 
     /// read little-endian word located at *addr*
-    fn read_lw(&self, addr: u32) -> u32 {
+    fn read_lw(&self, addr: usize) -> u32 {
         (self.memory[addr] as u32) |
         (self.memory[addr+1] as u32) << 8 |
         (self.memory[addr+2] as u32) << 16 |
@@ -37,7 +37,7 @@ impl VectorMemory {
     }
 
     /// write big-endian word at *addr*
-    fn write_bw(&mut self, addr: u32, val: u32) {
+    fn write_bw(&mut self, addr: usize, val: u32) {
         self.memory[addr] = (val >> 24) as u8;
         self.memory[addr+1] = (val >> 16) as u8;
         self.memory[addr+2] = (val >> 8) as u8;
@@ -45,7 +45,7 @@ impl VectorMemory {
     }
     
     /// write little-endian word at *addr*
-    fn write_lw(&mut self, addr: u32, val: u32) {
+    fn write_lw(&mut self, addr: usize, val: u32) {
         self.memory[addr] = val as u8;
         self.memory[addr+1] = (val >> 8) as u8;
         self.memory[addr+2] = (val >> 16) as u8;
@@ -53,24 +53,24 @@ impl VectorMemory {
     }
 
     /// read an instruction located at addr
-    pub fn write_inst(&mut self, addr: u32, inst: u32) {
+    pub fn write_inst(&mut self, addr: usize, inst: u32) {
         self.write_bw(addr, inst);
     }
 }
 
 impl Memory for VectorMemory {
     /// read an instruction located at *addr*
-    fn read_inst(&self, addr: u32) -> u32 {
+    fn read_inst(&self, addr: usize) -> u32 {
         self.read_bw(addr)
     }
 
     /// read word located at *addr*
-    fn read_word(&self, addr: u32) -> u32 {
+    fn read_word(&self, addr: usize) -> u32 {
         self.read_lw(addr)
     }
 
     /// write word at *addr*
-    fn write_word(&mut self, addr: u32, data: u32) {
+    fn write_word(&mut self, addr: usize, data: u32) {
         self.write_lw(addr, data);
     }
 }
